+++ application/modules/journal/controllers/SubmitController.php
@@ -12,21 +12,42 @@ class SubmitController extends DefaultController
     public function indexAction(): void
     {
 
+        $isFromZSubmit = false;
         $default = [];
         $settings = Zend_Registry::get('reviewSettings');
 
         /** @var Zend_Controller_Request_Http $request */
         $request = $this->getRequest();
-        $zIdentifier = filter_var($request->getParam('id'), \FILTER_VALIDATE_INT);
-        $requestedFrom = 'zsubmit';
-        $pattern = '#^' . $requestedFrom . '[-a-z]*.' . DOMAIN . '$#';
 
-        $isFromZSubmit = preg_match($pattern, $request->getServer('REMOTE_HOST'), $matches) && $zIdentifier &&
-            in_array(Episciences_Repositories::ZENODO_REPO_ID , $settings['repositories'], true);
+        if ($request->isPost()) {
 
-        if ($isFromZSubmit) {
-            $default ['repoId'] = Episciences_Repositories::ZENODO_REPO_ID;
-            $default ['docId'] = $zIdentifier;
+            $post = $request->getPost()['episciences_form'];
+
+            $repoId = $post['repoid'];
+            $zIdentifier = Episciences_Repositories::callHook('hookCleanIdentifiers', ['id' => $post['doi_show'], 'repoId' => $repoId])['identifier'];
+            $zConceptIdentifier = $request->getPost('ci');
+
+            $isFromZSubmit = $zIdentifier && $zConceptIdentifier && in_array($repoId, $settings['repositories'], true);
+
+            if( $isFromZSubmit){
+
+                $paper = Episciences_PapersManager::findByIdentifier($zConceptIdentifier);
+
+                $isFirstSubmission = !$paper || (
+                        $paper->getConcept_identifier() === $zConceptIdentifier &&
+                        in_array($paper->getStatus(), [Episciences_Paper::STATUS_SUBMITTED, Episciences_Paper::STATUS_OK_FOR_REVIEWING, Episciences_Paper::STATUS_REFUSED], true)
+                    );
+
+                if (!$isFirstSubmission) {
+                    $this->redirect($this->view->url(['controller' => 'paper', 'action' => 'view', 'id' => $paper->getDocid()], null, true));
+                    return;
+                }
+
+                $default ['repoId'] = Episciences_Repositories::ZENODO_REPO_ID;
+                $default ['docId'] = $zIdentifier;
+
+            }
+
         }
 
         $submit = new Episciences_Submit();